Add main.py to code.py
